Remove NaN value from set of scientific names
The test_create_species_dirs_all was returning True
even though it should have been False due to the fact
that the .info directory was not being removed from the list
of local directories.  This is because there was a NaN value
in the set of scientific names.

Also cleaned up the code that removes the .info directory from
the list of local directories.